requirements.txt
fastapi==0.122.0
uvicorn[standard]==0.38.0
joblib==1.5.2
numpy==2.0.2
pandas==2.2.2
scikit-learn==1.6.1
sklearn-pandas==2.2.0
python-multipart==0.0.20
huggingface-hub==0.26.2
python-dotenv==1.2.1
xgboost==3.1.2
pytest==9.0.1